requirements.txt
boto3>=1.34.34
botocore>=1.34.34
python-dotenv>=1.0.1
langchain-core>=0.1.27
langchain-community>=0.0.21
langchain-aws>=0.2.12
pandas>=2.2.0
matplotlib>=3.8.2